pyproject.toml
[project]
name = "employeerating-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery[redis]>=5.6.2",
    "fastapi>=0.129.0",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.0",
    "python-dotenv>=1.2.1",
    "redis>=6.4.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "uvicorn>=0.41.0",
]
